requirements.txt
annotated-types==0.6.0
blinker==1.6.2
charset-normalizer==2.0.12
click==8.1.3
dateparser==1.2.0
distro==1.8.0
Flask==2.3.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
itsdangerous==2.1.2
MarkupSafe==2.1.2
openai==1.3.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
pytz==2021.1
regex==2021.4.4
requests==2.26.0
ticktick-py==2.0.3
tqdm==4.66.1
tzlocal==5.2
urllib3==1.26.7
Werkzeug==2.3.3
